+++ Basil/include/Renderer/Renderer2D.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include "Renderer/OrthographicCamera.h"
+#include "Renderer/Texture2D.h"
 
 namespace Basil
 {
@@ -16,5 +17,7 @@ namespace Basil
 			static void endScene();
 			static void drawQuad(const glm::vec2& position, const glm::vec2& size, const glm::vec4& color);
 			static void drawQuad(const glm::vec3& position, const glm::vec2& size, const glm::vec4& color);
+			static void drawQuad(const glm::vec2& position, const glm::vec2& size, const Shared<Texture2D>& texture);
+			static void drawQuad(const glm::vec3& position, const glm::vec2& size, const Shared<Texture2D>& texture);
 	};
 }