+++ Basil/src/Platform/OpenGL/OpenGLRendererAPI.cpp
@@ -11,6 +11,7 @@ namespace Basil
 		// Enable texture blending
 		glEnable(GL_BLEND);
 		glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
+		glEnable(GL_DEPTH_TEST);
 	}
 
 	// Change the viewport