+++ Basil/src/Platform/OpenGL/OpenGLTexture2D.cpp
@@ -45,6 +45,8 @@ namespace Basil
 		// Set texture parameters
 		glTextureParameteri(rendererID, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 		glTextureParameteri(rendererID, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
+		glTextureParameteri(rendererID, GL_TEXTURE_WRAP_S, GL_REPEAT);
+		glTextureParameteri(rendererID, GL_TEXTURE_WRAP_T, GL_REPEAT);
 
 		// Upload texture
 		glTextureSubImage2D(rendererID, 0, 0, 0, this->width, this->height, dataFormat, GL_UNSIGNED_BYTE, data);