+++ Sandbox/include/Sandbox2D.h
@@ -19,5 +19,6 @@ class Sandbox2D : public Basil::Layer
 		Basil::OrthographicCameraController cameraController;
 		Basil::Shared<Basil::VertexArray> vao;
 		Basil::Shared<Basil::Shader> shader;
+		Basil::Shared<Basil::Texture2D> texture;
 		glm::vec4 squareColor = { 0.2f, 0.3f, 0.8f, 1.0f };
 };